% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract coefficients from a linreg model}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg model.}

\item{...}{Additional arguments.}
}
\value{
A named vector containing the regression coefficients.
}
\description{
Extract coefficients from a linreg model
}
