% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linreg Class}
\value{
Nothing.
}
\description{
This class performs linear regression and provides different methods to display the data. It also prints the plots for Residual vs. Fitted Values and Scale-Location
}
\details{
The algorithm used for this function can be found on 
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
\section{Fields}{

\describe{
\item{\code{coefficients}}{matrix. Regression Coefficients.}

\item{\code{fitted}}{matrix. Fitted Values.}

\item{\code{residuals}}{matrix. Residuals.}

\item{\code{degrees_of_freedom}}{numeric. Degrees of freedom.}

\item{\code{resvar}}{matrix. Residual Variance.}

\item{\code{coefficients_variance}}{matrix. Variance of the Regression Coefficients.}

\item{\code{formula}}{formula. The Formula for the Linear Regression.}

\item{\code{pvalues}}{vector. P-values.}

\item{\code{tvalues}}{vector. T-values for each Coefficient.}

\item{\code{data_name}}{character. The Given Data.}

\item{\code{data}}{data.frame. The given data frame.}

\item{\code{call_params}}{character. A pairlist of all active calls.}

\item{\code{outliers}}{vector. Outliers.}
}}


