% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Return predictions from a linreg model}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg model.}

\item{...}{Additional arguments.}
}
\value{
The fitted values from the model.
}
\description{
Return predictions from a linreg model
}
