% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{pred}
\alias{pred}
\title{Return predictions from a linreg model}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{A linreg model.}

\item{...}{Additional arguments.}
}
\value{
A vector containing the fitted values.
}
\description{
Return predictions from a linreg model
}
