% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.linreg}
\alias{print.linreg}
\title{Print a linreg model}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg model.}

\item{...}{Additional arguments.}
}
\value{
The model object, invisibly.
}
\description{
Print a linreg model
}
