% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Function to return residual values}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{other arguments}
}
\value{
Returns residual values
}
\description{
Function to return residual values
}
\examples{
data(iris)
mod_object <- linreg(formula = Petal.Length ~ Species, data = iris)
resid(mod_object)
}
