% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summarize a linreg model}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg model.}

\item{...}{Additional arguments.}
}
\value{
A summary of the regression model.
}
\description{
Summarize a linreg model
}
