% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Function to return the coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Other arguments}
}
\value{
Returns a vector containing the calculated coefficient values
}
\description{
Function to return the coefficients
}
\examples{
data(iris)
mod_object <- linreg(formula = Petal.Length ~ Species, data = iris)
coef(mod_object)
}
