% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Function that creates an S3 Class of class linreg}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{a formula object}

\item{data}{A data frame}
}
\value{
A class of type linreg with specific attributes and functions
}
\description{
Function that creates an S3 Class of class linreg
}
\examples{
data(iris)
linreg(formula = Petal.Length ~ Species, data = iris)


}
