% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linerg.R
\name{linreg}
\alias{linreg}
\title{Linear regression model}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula describing the regression model.}

\item{data}{A data frame containing the variables in the formula.}
}
\value{
A linreg object containing the coefficients, fitted values,
residuals, standard errors, t-values, p-values and other model information.
}
\description{
Fits a linear regression model using the least squares method.
}
