% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Function to Plot Results}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list contains ggplots.
}
\description{
Function to Plot Results
}
\examples{
data(iris)
mod_object <- linreg(formula = Petal.Length ~ Species, data = iris)
plot(mod_object)
}
