% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot a linreg model}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg model.}

\item{...}{Additional arguments.}
}
\value{
A list containing two ggplot objects.
}
\description{
Plot a linreg model
}
