% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Function to Return Predicted Values}
\usage{
\method{pred}{linreg}(obj)
}
\arguments{
\item{obj}{An object of class linreg}
}
\value{
Function returns a numeric vector of all the predicted values
}
\description{
Function to Return Predicted Values
}
\examples{
data(iris)
mod_object <- linreg(formula = Petal.Length ~ Species, data = iris)
pred(mod_object)
}
