% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{pred}
\alias{pred}
\title{Prediction function for linreg models}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments.}
}
\value{
Predicted values.
}
\description{
Prediction function for linreg models
}
