% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Custom Print Function}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An list object containing all the relevant information calculated by the function linreg or lr}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns coefficients
}
\description{
This is a custom print function that prints the coefficients
}
\examples{
data(iris)
mod_object <- linreg(formula = Petal.Length ~ Species, data = iris)
print(mod_object)
}
