% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Extract residuals from a linreg model}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg model.}

\item{...}{Additional arguments.}
}
\value{
The model residuals.
}
\description{
Extract residuals from a linreg model
}
