% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Function to print Summary of Results}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{A object of class linreg}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns a data frame consisting of results.
}
\description{
Function to print Summary of Results
}
\examples{
data(iris)
mod_object <- linreg(formula = Petal.Length ~ Species, data = iris)
summary(mod_object)
}
